[AAE-10390] Storybook add missed code
