--- lib/core/src/lib/dialogs/download-zip/mocks/stub.data.ts
@@ -1,33 +1,0 @@
-/*!
- * @license
- * Copyright 2019 Alfresco Software, Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { DownloadEntry } from '@alfresco/js-api';
-
-export const zipNode = {
-    entry: {
-        name: 'files.zip',
-        contentUrl: './assets/files.zip',
-        id: 'files_in_zip'
-    }
-};
-
-export const downloadEntry: DownloadEntry = {
-    entry: {
-        id: 'entryId',
-        status: 'DONE'
-    }
-};
